reorg remaining patches into Custom Features
