+++ SolastaCommunityExpansion/Patches/CustomFeatures/CharacterBuildingManagerPatcher.cs
@@ -1,13 +1,11 @@
 ﻿using HarmonyLib;
 using SolastaCommunityExpansion.CustomFeatureDefinitions;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 
 namespace SolastaCommunityExpansion.Patches.CustomFeatures
 {
-    //
-    // this patch shouldn't be protected
-    //
     [HarmonyPatch(typeof(CharacterBuildingManager), "FinalizeCharacter")]
     [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
     internal static class CharacterBuildingManager_FinalizeCharacter
@@ -24,6 +22,19 @@ namespace SolastaCommunityExpansion.Patches.CustomFeatures
                     featureDefinitions.RemoveAll(x => x == removeGrantedFeature.FeatureToRemove);
                 }
             }
+        }
+    }
+
+    [HarmonyPatch(typeof(CharacterBuildingManager), "GrantFeatures")]
+    [SuppressMessage("Minor Code Smell", "S101:Types should be named in PascalCase", Justification = "Patch")]
+    internal static class CharacterBuildingManager_GrantFeatures
+    {
+        /**
+         * When a character is being granted features, this patch will apply the effect of custom features.
+         */
+        internal static void Prefix(CharacterBuildingManager __instance, List<FeatureDefinition> grantedFeatures)
+        {
+            Models.CustomFeaturesContext.RecursiveGrantCustomFeatures(__instance.HeroCharacter, grantedFeatures);
         }
     }
 }
